a little refactoring separate calculateExpression function into separate js module file
