+++ index.html
@@ -81,6 +81,7 @@
         </ul>
       </section>
     </div>
-    <script src="./src/main.js"></script>
+    <script type="module" src="./src/main.js"></script>
+    <script type="module" src="./src/calculateExpression.js"></script>
   </body>
 </html>
